Exp 04

+++ Form_4.MapEvents.cs
@@ -5,10 +5,9 @@ using ESRI.ArcGIS.Geometry;
 using System;
 using System.Linq;
 using System.Windows.Forms;
-using Lab03_4.MyForms.FeatureManagement.Services;
 using ESRI.ArcGIS.Geodatabase;
 
-namespace Lab03_4
+namespace Lab04_4
 {
     public partial class Form_4 : Form
     {
@@ -102,6 +101,12 @@ namespace Lab03_4
                     features = SelectAllFeatures(geometry);
                 }
                 if (features is null) return;
+
+                // 弹出提示确认
+                DialogResult dialogResult =
+                MessageBox.Show("是否要删除选中要素？该操作不可撤回", "删除要素", MessageBoxButtons.OKCancel);
+                if (dialogResult != DialogResult.OK) return;
+
                 DeleteFeatures(features);
                 axMap.Map.ClearSelection();
             }
@@ -113,6 +118,33 @@ namespace Lab03_4
                 IFeature feature = SelectFirstFeature(geometry);
                 if (feature is null) return;
                 IdentifyFeature(feature);
+            }
+            else if (mapOperation == MapOperationType.IntepolateElevation)
+            {
+                IPoint clickPoint = sketcher.Sketch(axMap, e) as IPoint;
+                IntepolateElevation(clickPoint);
+                mapOperation = MapOperationType.Default;
+            }
+            else if (mapOperation == MapOperationType.ElementQuery)
+            {
+                IPoint clickPoint = sketcher.Sketch(axMap, e) as IPoint;
+                SpatialQueryService.QueryElement(clickPoint);
+            }
+            else if (mapOperation == MapOperationType.DrawPolyline)
+            {
+                // if (e.button == 1)
+                // {
+                //    IPoint clickPoint = sketcher.Sketch(axMap, e) as IPoint;
+                //    SpatialQueryService.AddPoint(clickPoint);
+                // }
+                // else if (e.button == 2)  // 右键结束
+                // {
+                //    SpatialQueryService.FinishDrawing();
+                //    mapOperation = MapOperationType.Default;
+                // }
+                IGeometry geometry = sketcher.Sketch(axMap, e);
+                SpatialQueryService.FinishDrawingLine(geometry as IPolyline);
+                mapOperation = MapOperationType.Default;
             }
         }
